Moved more logic into ResolvedByFilter
